Define ResultSet interface for ODBC abstraction
This defines the ResultSet interface for the ODBC abstraction layer.

Based on discussions between me and James we decided to follow a pattern
similar to ODBC's `SQLBindCol` and `SQLFetch`, but for multiple rows
(batch) in order to keep the interface efficient by reducing the virtual
method calls and also hope to facilitate when implement bulk methods
(ie. bulk conversion)

Change-Id: I5e4c9de4cbd012f3745ae1e7600256dedb8f3c39