Fix SQLGetData not switching arrays when iterating chunks
SQLGetData was re-using the same array from the first chunk in
a FlightStream, rather than rebuilding its accessor with the next
array in the chunk.

This would cause data to repeat itself over and over, and also
crash when iterating over multi-chunk data because the size of
all chunks isn't the same (so there'd be an array out of bounds
error).

Change-Id: I2413fae26f64caa41ac122861aa5fe810a94d5cd