Fix format for SQL_DBMS_VER
Change SQL_DBMS_VER to use the format ##.##.####<custom-info>
to match the ODBC spec.

Interpret FLIGHT_SQL_SERVER_VERSION when it has missing or
invalid values in the version number to match the format above. (Fill out dummy
values when we see an illegal sequence then paste the illegal sequence at the end).

Change-Id: I93a4d7c6d6db7de9c013705b55c586138559f55c