Add `UseExtendedFlightSQLBuffer` flag
When set to `true` (by default it's `false`) it will enable an extended
buffer to keep retrieving `FlightStreamChunk`s in a background thread
while the client can't consume the early instances.

This should provide a workaround for slow clients that take a lot of time
to process query results, potentially having the Flight connection severed
because the lack of activity.

Change-Id: I0275dbb70d2971b9b9b20d1598b19b067f42f987